feat: replace buildPgDatabase with buildPgDatabaseWithMigrations
